@@ -4,37 +4,35 @@ version: 2.0.0-alpha-01
 homepage: https://github.com/zubairehman/Flogs
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
   # An helper to easily exposes a value using InheritedWidget without having to write one.
-  sembast: ^3.0.0+4
-  path_provider: ^2.0.1
+  sembast: ^3.1.1+1
+  path_provider: ^2.0.8
 
   # This package provides internationalization and localization facilities, including message
   # translation, plurals and genders, date/number formatting and parsing, and bidirectional text.
   intl: ^0.17.0
 
   # XXTEA is a fast and secure encryption algorithm. This is a XXTEA library for Dart.
-  xxtea: ^2.0.3
+  xxtea: ^2.1.0
 
   pointycastle: ^3.0.1
 
   # Tuple data structure
   # tuple: ^1.0.2
 
-  meta: ^1.1.8
-  path: ^1.6.4
-  stack_trace: ^1.9.3
+  meta: ^1.7.0
+  path: ^1.8.0
+  stack_trace: ^1.10.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-  effective_dart: ^1.2.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
