@@ -1,27 +1,26 @@
 import 'dart:convert';
 
-import 'package:meta/meta.dart';
 import 'package:sembast/sembast.dart';
 import 'package:xxtea/xxtea.dart';
 
-class _XXTeaEncoder extends Converter<Map<String, dynamic>, String> {
+class _XXTeaEncoder extends Converter<Object?, String> {
   final String key;
 
   _XXTeaEncoder(this.key);
 
   @override
-  String convert(Map<String, dynamic> input) =>
-      xxtea.encryptToString(json.encode(input), key);
+  String convert(Object? input) =>
+      xxtea.encryptToString(json.encode(input), key)!;
 }
 
-class _XXTeaDecoder extends Converter<String, Map<String, dynamic>> {
+class _XXTeaDecoder extends Converter<String, Object?> {
   final String key;
 
   _XXTeaDecoder(this.key);
 
   @override
   Map<String, dynamic> convert(String input) {
-    var result = json.decode(xxtea.decryptToString(input, key));
+    var result = json.decode(xxtea.decryptToString(input, key)!);
     if (result is Map) {
       return result.cast<String, dynamic>();
     }
@@ -31,9 +30,9 @@ class _XXTeaDecoder extends Converter<String, Map<String, dynamic>> {
 
 /// Simple encryption codec using xxtea
 /// It requires a password to encrypt/decrypt the data
-class _XXTeaCodec extends Codec<Map<String, dynamic>, String> {
-  _XXTeaEncoder _encoder;
-  _XXTeaDecoder _decoder;
+class _XXTeaCodec extends Codec<Object?, String> {
+  late _XXTeaEncoder _encoder;
+  late _XXTeaDecoder _decoder;
 
   /// A non null [password] to use for the encryption/decryption
   _XXTeaCodec(String password) {
@@ -42,10 +41,10 @@ class _XXTeaCodec extends Codec<Map<String, dynamic>, String> {
   }
 
   @override
-  Converter<String, Map<String, dynamic>> get decoder => _decoder;
+  Converter<String, Object?> get decoder => _decoder;
 
   @override
-  Converter<Map<String, dynamic>, String> get encoder => _encoder;
+  Converter<Object?, String> get encoder => _encoder;
 }
 
 /// Create a codec to use when opening an encrypted sembast database
@@ -60,5 +59,5 @@ class _XXTeaCodec extends Codec<Map<String, dynamic>, String> {
 ///
 /// // ...your database is ready to use as encrypted
 /// ```
-SembastCodec getXXTeaSembastCodec({@required String password}) =>
+SembastCodec getXXTeaSembastCodec({required String password}) =>
     SembastCodec(signature: 'xxtea', codec: _XXTeaCodec(password));
